# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.1",
    "title": "Ecommerce backend API",
    "description": "Ecommerce backend API with Express"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/orders/": {
      "get": {
        "tags": [
          "orders"
        ],
        "description": "Fetch all orders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "all orders",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 132
                  },
                  "title": {
                    "type": "string",
                    "example": "PEGASUS 33"
                  },
                  "description": {
                    "type": "string",
                    "example": "The Zoom Pegasus ...."
                  },
                  "price": {
                    "type": "number",
                    "example": 59.99
                  },
                  "username": {
                    "type": "string",
                    "example": "john"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No orders found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "No orders found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": [
          "orders"
        ],
        "description": "Get Single Order",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Order ID"
          }
        ],
        "responses": {
          "200": {
            "description": "all orders",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 174
                  },
                  "title": {
                    "type": "string",
                    "example": "PEGASUS 33 Running Shoes For Men"
                  },
                  "description": {
                    "type": "string",
                    "example": "The Nike Zoom Pegasus Turbo 2 is updated with a feather-light upper, while innovative foam brings revolutionary responsiveness to your long-distance training"
                  },
                  "price": {
                    "type": "number",
                    "example": 59.99
                  },
                  "image": {
                    "type": "string",
                    "example": "https://i.pinimg.com/originals/43/40/8e/43408ee5a8d234752ecf80bbc3832e65.jpg"
                  },
                  "quantityOrdered": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/orders/new": {
      "post": {
        "tags": [
          "orders"
        ],
        "description": "Place New Order",
        "parameters": [
          {
            "name": "prodId",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "products": {
                  "example": "any"
                }
              }
            },
            "description": "user & products"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "description": "LOGIN",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "user for login",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@email.com"
                },
                "password": {
                  "type": "string",
                  "example": "************"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user log",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string",
                  "example": "****************************"
                },
                "auth": {
                  "type": "boolean",
                  "example": true
                },
                "email": {
                  "type": "string",
                  "example": "user@mail.com"
                },
                "firstName": {
                  "type": "string",
                  "example": "firstName"
                },
                "lastName": {
                  "type": "string",
                  "example": "lastName"
                },
                "photoUrl": {
                  "type": "string",
                  "example": "photoUrl.ulr.com"
                },
                "id": {
                  "type": "number",
                  "example": 0
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "authentication"
        ],
        "description": "REGISTER",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "user for login",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@email.com"
                },
                "password": {
                  "type": "string",
                  "example": "***********"
                },
                "fname": {
                  "type": "string",
                  "example": "Leo"
                },
                "lname": {
                  "type": "string",
                  "example": "Edward"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User successfully reg.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Registration successful."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "433": {
            "description": "User reg error"
          },
          "501": {
            "description": "User reg failed",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Registration failed."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}